[feature]添加debugActivity
