+++ module_main/build.gradle
@@ -5,4 +5,6 @@ apply plugin: 'kotlin-android-extensions'
 // 这里只添加确认模块独用库，添加请之前全局搜索，是否已经依赖
 // 公用库请不要添加到这里
 dependencies {
+implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.1'
 }
