+++ module_main/src/main/AndroidManifest.xml
@@ -3,22 +3,17 @@
     package="com.mredrock.cyxbs.main">
 
     <application android:theme="@style/AppTheme">
-
-
+        <activity android:name=".debug.DebugActivity"></activity>
         <activity
             android:name=".ui.MainActivity"
             android:exported="true"
             android:launchMode="singleTop"
             android:theme="@style/SplashTheme">
-
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
-
-
             </intent-filter>
-
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
 
@@ -29,15 +24,13 @@
                     android:host="qa"
                     android:scheme="cyxbsmobile" />
 
-                <!--<a href="[scheme]://[host]/[path]?[query]">启动应用程序</a>-->
+                <!-- <a href="[scheme]://[host]/[path]?[query]">启动应用程序</a> -->
             </intent-filter>
 
             <meta-data
                 android:name="android.app.shortcuts"
                 android:resource="@xml/main_shortcut" />
-
         </activity>
-
         <activity
             android:name=".ui.LoginActivity"
             android:launchMode="singleTask" />
